+++ src/explorer/character.hpp
@@ -14,8 +14,8 @@ public:
     virtual void draw(sf::RenderTarget& target, sf::RenderStates states) const override;
 
     int get_max_speed() const;
-    sf::Vector2i get_current_velocity() const;
-    void set_current_velocity(sf::Vector2i updated_velocity);
+    sf::Vector2f get_current_velocity() const;
+    void set_current_velocity(sf::Vector2f updated_velocity);
     sf::Vector2i get_current_acc() const;
     void set_current_acc(sf::Vector2i updated_acc);
 
@@ -27,11 +27,11 @@ public:
 protected:
     virtual void set_sprite(sf::Vector2i const& start_pose) = 0;
 
-    int _max_speed;
+    float _max_speed;
     sf::Sprite _sprite;
     std::string _name;
     sf::Texture _texture;
-    sf::Vector2i _current_velocity;
+    sf::Vector2f _current_velocity;
     sf::Vector2i _current_acc;
 };
